currency_rate/requirements.txt
aio-pika==9.3.0
aiohttp==3.8.6
aiormq==6.7.7
aiosignal==1.3.1
alembic==1.12.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
binance-connector==3.5.0
certifi==2023.7.22
charset-normalizer==3.3.2
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
pamqp==3.2.1
pycryptodome==3.19.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.0.7
websocket-client==1.6.4
yarl==1.9.2
